+++ src/streamindexer/streamendanalyzer.cpp
@@ -3,9 +3,9 @@
 
 char
 StreamEndAnalyzer::testStream(InputStream *in) {
-    size_t testsize = 1;
+    int32_t testsize = 1;
     const char *dummyptr;
-    size_t dummy;
+    int32_t dummy;
     char r = in->mark(testsize);
     if (r) {
         return r;
